Gate various functions on required binaries or platforms.

+++ fish/.config/fish/functions/flush_dns.fish
@@ -1,3 +1,5 @@
 function flush_dns
-  sudo killall -HUP mDNSResponder
+  if is_mac
+    sudo killall -HUP mDNSResponder
+  end
 end
